+++ components/log-viewer-webui/client/src/AntdApp.tsx
@@ -10,6 +10,7 @@ import THEME_CONFIG from "./theme";
 
 import "@ant-design/v5-patch-for-react-19";
 
+
 /**
  * Renders Web UI app.
  *
