+++ components/log-viewer-webui/client/src/pages/IngestPage/Details/index.tsx
@@ -7,23 +7,16 @@ import {
     SETTINGS_STORAGE_ENGINE,
 } from "../../../config";
 import {fetchDatasetNames} from "../../SearchPage/SearchControls/Dataset/sql";
-import useIngestStatsStore from "../ingestStatsStore";
 import Files from "./Files";
 import styles from "./index.module.css";
 import Messages from "./Messages";
 import {
     buildMultiDatasetDetailsSql,
+    DETAILS_DEFAULT,
     DetailsItem,
 } from "./sql";
 import TimeRange from "./TimeRange";
 
-
-const DETAILS_DEFAULT: DetailsItem = {
-    begin_timestamp: null,
-    end_timestamp: null,
-    num_files: 0,
-    num_messages: 0,
-};
 
 /**
  * Renders grid with compression details.
@@ -31,20 +24,18 @@ const DETAILS_DEFAULT: DetailsItem = {
  * @return
  */
 const Details = () => {
-    const {refreshInterval} = useIngestStatsStore();
 
     const {data: datasetNames, isSuccess: isSuccessDatasetNames} = useQuery({
         queryKey: ["datasets"],
         queryFn: fetchDatasetNames,
-        staleTime: refreshInterval,
     });
 
-    const {data: details, isPending} = useQuery({
+    const {data, isPending} = useQuery({
         queryKey: ["aggregate-stats",
             datasetNames],
         queryFn: async () => {
             if (false === isSuccessDatasetNames) {
-                throw new Error("Dataset names are not available");
+                throw new Error("Dataset names are not available.");
             }
             if (0 === datasetNames.length) {
                 return DETAILS_DEFAULT;
@@ -53,7 +44,7 @@ const Details = () => {
             const resp = await querySql<DetailsItem[]>(sql);
             const [detailsResult] = resp.data;
             if ("undefined" === typeof detailsResult) {
-                throw new Error("Details response is undefined");
+                throw new Error("Details result does not contain data.");
             }
 
             return detailsResult;
@@ -61,21 +52,23 @@ const Details = () => {
         enabled: isSuccessDatasetNames,
     });
 
+    const details = data ?? DETAILS_DEFAULT;
+
     if (CLP_STORAGE_ENGINES.CLP === SETTINGS_STORAGE_ENGINE) {
         return (
             <div className={styles["detailsGrid"]}>
                 <div className={styles["timeRange"]}>
                     <TimeRange
-                        beginDate={dayjs.utc(details?.begin_timestamp ?? null)}
-                        endDate={dayjs.utc(details?.end_timestamp ?? null)}
+                        beginDate={dayjs.utc(details.begin_timestamp)}
+                        endDate={dayjs.utc(details.end_timestamp)}
                         isLoading={isPending}/>
                 </div>
                 <Messages
                     isLoading={isPending}
-                    numMessages={details?.num_messages ?? 0}/>
+                    numMessages={details.num_messages}/>
                 <Files
                     isLoading={isPending}
-                    numFiles={details?.num_files ?? 0}/>
+                    numFiles={details.num_files}/>
             </div>
         );
     }
@@ -83,8 +76,8 @@ const Details = () => {
     return (
         <div>
             <TimeRange
-                beginDate={dayjs.utc(details?.begin_timestamp ?? null)}
-                endDate={dayjs.utc(details?.end_timestamp ?? null)}
+                beginDate={dayjs.utc(details.begin_timestamp)}
+                endDate={dayjs.utc(details.end_timestamp)}
                 isLoading={isPending}/>
         </div>
     );
